Insert and remove nodes with child modifications for Android Views
