+++ redwood-layout-view/src/main/kotlin/app/cash/redwood/layout/view/ViewMeasureCallback.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2023 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.redwood.layout.view
+
+import android.view.View
+import app.cash.redwood.yoga.MeasureCallback
+import app.cash.redwood.yoga.MeasureMode
+import app.cash.redwood.yoga.Node
+import app.cash.redwood.yoga.Size
+import kotlin.math.roundToInt
+
+internal class ViewMeasureCallback(val view: View) : MeasureCallback {
+  override fun measure(
+    node: Node,
+    width: Float,
+    widthMode: MeasureMode,
+    height: Float,
+    heightMode: MeasureMode,
+  ): Size {
+    val safeWidth = if (width.isFinite()) width.roundToInt() else 0
+    val safeHeight = if (height.isFinite()) height.roundToInt() else 0
+    val widthSpec = View.MeasureSpec.makeMeasureSpec(safeWidth, widthMode.toAndroid())
+    val heightSpec = View.MeasureSpec.makeMeasureSpec(safeHeight, heightMode.toAndroid())
+    view.measure(widthSpec, heightSpec)
+    return Size(view.measuredWidth.toFloat(), view.measuredHeight.toFloat())
+  }
+}
